// src/component/Facetec/declarations/FaceTecPublicApi.d.ts
export declare enum FaceTecIDScanStatus {
  /**
    The ID Scan was successful.
   */
  Success = 0,
  /**
     The ID Scan was not successful
    */
  Unsuccess = 1,
  /**
     User cancelled ID Scan
    */
  UserCancelled = 2,
  /**
     Timeout during ID Scan
    */
  TimedOut = 3,
  /**
     Context Switch during ID Scan
    */
  ContextSwitch = 4,
  /**
     Error setting up the ID Scan Camera
    */
  CameraError = 5,
  /**
     Camera Permissions were not enabled
    */
  CameraNotEnabled = 6,
  /**
     ID Scan was skipped.
    */
  Skipped = 7,
  /**
     ID Scan was cancelled because device is in landscape mode.
    */
  LandscapeModeNotAllowed = 8,
}
/** ID Scan Result object */
export interface FaceTecIDScanResult {
  status: FaceTecIDScanStatus;
  idScan: string | null;
  frontImages: string[];
  backImages: string[];
  sessionId: string | null;
  isCompletelyDone: boolean;
}
/** Callback functions for FaceTecIDScanProcessor */
export declare class FaceTecIDScanResultCallback {
  cancel: () => void;
  uploadProgress: (uploadedPercent: number) => void;
  uploadMessageOverride: (uploadMessageOverride: string) => void;
  proceedToNextStep: (
    scanResultBlob: string,
    idScanNextStep?: FaceTecIDScanNextStep
  ) => void;
}
/** Abstract class for developer to override for processing FaceTec ID Scans. */
export declare abstract class FaceTecIDScanProcessor {
  abstract onFaceTecSDKCompletelyDone: () => void;
  abstract processIDScanResultWhileFaceTecSDKWaits: (
    idScanResult: FaceTecIDScanResult,
    idScanResultCallback: FaceTecIDScanResultCallback
  ) => void;
}
// ...
export declare enum FaceTecIDScanNextStep {
  /**
   * Start ID Scan process with showing the Selection Screen.
   * This is default behavior.
   */
  SelectionScreen = 0,
  /**
   * Skip the entire ID Scan process, exiting from the FaceTec Browser SDK interface after a successful Session.
   */
  Skip = 1,
}